split rocketchat services into files
